fix: 폰트 적용 오류 수정
