+++ src/index.css
@@ -1,6 +1,7 @@
 @font-face {
   font-family: "Spoqa Han Sans Neo";
-  src: url("./assets/fonts/SpoqaHanSansNeo-Light.woff2") format("woff2"),
+  src: url("./assets/fonts/SpoqaHanSansNeo-Light.woff2")
+      format("woff2"),
     url("./assets/fonts/SpoqaHanSansNeo-Light.woff") format("woff");
   font-weight: 300;
   font-style: normal;
@@ -9,7 +10,8 @@
 
 @font-face {
   font-family: "Spoqa Han Sans Neo";
-  src: url("./assets/fonts/SpoqaHanSansNeo-Medium.woff2") format("woff2"),
+  src: url("./assets/fonts/SpoqaHanSansNeo-Medium.woff2")
+      format("woff2"),
     url("./assets/fonts/SpoqaHanSansNeo-Medium.woff") format("woff");
   font-weight: 500;
   font-style: normal;
@@ -18,7 +20,8 @@
 
 @font-face {
   font-family: "Spoqa Han Sans Neo";
-  src: url("./assets/fonts/SpoqaHanSansNeo-Bold.woff2") format("woff2"),
+  src: url("./assets/fonts/SpoqaHanSansNeo-Bold.woff2")
+      format("woff2"),
     url("./assets/fonts/SpoqaHanSansNeo-Bold.woff") format("woff");
   font-weight: bold;
   font-style: normal;
@@ -27,7 +30,8 @@
 
 @font-face {
   font-family: "Spoqa Han Sans Neo";
-  src: url("./assets/fonts/SpoqaHanSansNeo-Regular.woff2") format("woff2"),
+  src: url("./assets/fonts/SpoqaHanSansNeo-Regular.woff2")
+      format("woff2"),
     url("./assets/fonts/SpoqaHanSansNeo-Regular.woff") format("woff");
   font-weight: normal;
   font-style: normal;
@@ -36,7 +40,8 @@
 
 @font-face {
   font-family: "Spoqa Han Sans Neo";
-  src: url("./assets/fonts/SpoqaHanSansNeo-Thin.woff2") format("woff2"),
+  src: url("./assets/fonts/SpoqaHanSansNeo-Thin.woff2")
+      format("woff2"),
     url("./assets/fonts/SpoqaHanSansNeo-Thin.woff") format("woff");
   font-weight: 100;
   font-style: normal;
